Use sorted() to sort keys in a dict on Python 3
In Python 2 we used to sort the keys in a dictionary by retrieving a
list of keys and sorting them in place. The dict.keys() method now
returns a view, so we no longer have a sort() method to call.

Luckily we can use sorted() to return a sorted list of keys, which works
in Python 2.5+.

https://docs.python.org/3.1/whatsnew/3.0.html#views-and-iterators-instead-of-lists